Calculate and update mhalo
Previously, the total halo mass mhalo was never updated, but instead stayed untouched at its inital value. Not it's being integrated from its initial value onwards, using MIR * timestep to add and pre-integrated mout * HLF to decrease it; HLF is the newly introduced halo loss fraction: the fraction of mout gas that is not only ejected from the galaxy by feedback, but also from the whole halo and is, for our purposes, lost from the (Galaxy) system.

We might want to add that ejected (mout * HLF) to the Environment instance mgas in the future, although it has mgas = np.inf at the moment.

Related issue #3
Resolves issue #1