Adapt SFR, MLR, GCR for 'intuitive' integrator
In further pursuit of #7:

Added alternative compute_XXX methods for
  * SFR (Star Formation Rate)
  * MLR (Mass Loss Rate)
  * GCR (Gas Change Rate)
in accordance with the new, more 'intuitive' evolve/integrator. In the
course of that renamed the existing compute_XXX methods. Either version
now includes a descriptive addition to the method name to specify which
version it is, e.g. compute_SFR_from_SFE().

The different versions of compute_XXX are called by providing the
update_XXX methods with an according 'mode' keyword, e.g.
mode="from SFE". The calls in reference_evolve() and intuitive_evolve()
have been adapted accordingly so that the appropriate compute_XXX
method in each case is called.

Related to issue #7