Adapt mstar, mout, mgas for 'intuitive' integrator
In further pursuit of #7:

Added alternative compute_XXX methods for
  * mstar (Stellar Mass)
  * mgas (Gas Mass In Reservoir)
Did not need to add an alternative for
  * mout (Mass Lost From Galaxy)
In accordance with the new, more 'intuitive' evolve/integrator. In the
course of that renamed the existing compute_XXX methods. Either version
now includes a descriptive addition to the method name to specify which
version it is, e.g. compute_mstar_from_unreduced_SFR().

The different versions of compute_XXX are called by providing the
update_XXX methods with an according 'mode' keyword, e.g.
mode="from reduced SFR". The calls in reference_evolve() and
intuitive_evolve() have been adapted accordingly so that the appropriate
compute_XXX method in each case is called.

Also: fixed issue #8 where the integration/iterative computation of
mstar, mout and mgas was broken and only returned/saved the incremental
change w.r.t. the last time step instead of adding that to the previous
value.
  * Pre fix: mstar = SFR * timestep
  * Post fix: mstar = mstar + (SFR * timestep)

Related to issue #7
Fixed issue #8