Adapt fstar, fout, fgas for 'intuitive' integrator
In further pursuit of #7:

Added alternative compute_XXX methods for
  * fstar (Fraction of Accreted Gas into Stellar Mass)
  * fout (Fraction of Accreted Gas ejected out of Galaxy)
  * fgas (Fraction of Accreted Gas into Reservoir)
in accordance with the new, more 'intuitive' evolve/integrator. In the
course of that renamed the existing compute_XXX methods. Either version
now includes a descriptive addition to the method name to specify which
version it is, e.g. compute_fstar_beforehand(). The new compute_XXX
methods compare the actual increase in mstar, mgas, and mout with the
actual increase in gross accreted mass onto the galaxy. This can be used
to e.g. later-on compare to the 'reference' model (Lilly+13) that
heavily relies on these factors, albeit computed beforehand and
differently. NB: fstar + fout + fgas is not necessarily = 1 anymore, for
the 'intuitive' evolve.

The different versions of compute_XXX are called by providing the
update_XXX methods with an according 'mode' keyword, e.g.
mode="afterwards". The calls in reference_evolve() and
intuitive_evolve() have been adapted accordingly so that the appropriate
compute_XXX method in each case is called.

Also: removed previousy added alternative for
  * mgas (Gas Mass in Reservoir)
since they were redundant (as per issue #10).

Also: added a self.previous attribute that stores a snapshot, generated
by the make_snapshot() method, to allow for comparison with the previous
time step. All attributes except self.env are copied. At the moment this
might lead to recursion, because self.previous contains a self.previous
of its own, and ever so on. I circumvented this by setting self.previous
to None right before calling make_snapshot(), but this should be handled
better. See issue #11 for more details on possible solutions.

Related to issues #7, #11
Fixed issue #10