Add method to construct dict from attribute names
Added helper method `Galaxy.attr_to_dict(self, attr, default_attr)` to:
 (a) check whether attr was supplied at all. If not, then do use the
     default_attr list input from there on;
 (b) check whether attr was already supplied as a dict. If yes, then
     check that each key has a not None value, and if some do, then set
     them to the corresponding attribute values of the Galaxy object;
 (c) check if attr was supplied as list, and if yes, then construct a
     corresponding dict out of the entries as keys and the matching
     attribute values of the object as corresponding dict values.

Beware: lists are not checked for dimensionality (yet? not sure if this
is truly necessary), and *only* the keys are supposed to be supplied as
list entries. If there are corresponding values that you want to supply,
use a dict as input instead!

Not tested in practice yet.

Issue solved: #14