Add optional multiple timesteps per burn-in step
Based on issue #16, it is now possible to set cycle_steps to any integer
number for specifying how many time steps you want the system to evolve
during burn-in before the deviations (combined `div` and individual
`attr_div`) are calculated and before the assessment of whether the
burn-in has converged, is stable or has diverged happens, and before the
fixed quantities are reset.

A bit more information is printed to screen for the time being, and the
individual divs are now also returned in the form of `attr_div`
dictionary by `Galaxy.check_for_convstabdiv()` in addition to the
previously returned `div`, `converged`, `stable`, `diverged`. For
details on that, see previous descriptions.

Also, typos fixed and other minor corrections.

Resolves issue #16