Write own Snapshot class for outputs
Based on issue #20, I have written a Snapshot class to store output data
of any kind of object that currently returns a dict. These will mostly
(but not necessarily) be the upcoming `AstroObject` subclasses:
 * Environment
 * Halo
 * Galaxy

Since the original idea behind using a separate snapshot class instead
of plain Python dictionaries was to more easily recognise them (e.g. by
using `type(snapshotinstance)`, I created the `Snapshot` class as a
direct subclass of the standard Python `Dictionary` class.

This has very quickly become a problem, since I also started adding
additional attributes to it that are not part of regular dicts, namely:
 * `self.basename`
 * `self.snaptime`
 * `self.filepath`
 * `self.dirpath`

All of these properties do not contain simulation outputs, but some sort
of meta information that relates to when the snapshot was taken
(`snaptime`, used as a unique identifier) and under which name and path
the snapshot will be stored to disk (all others).

The problems now are that:
 * the dict-like key-value pairs are accessible in a dictionary-like
   fashion, e.g. stellar mass in snapshot `snp` as `snp['mstar']` ; at
   the same time, the meta-data is accessible through regular attribute
   syntax, e.g. `snp.filename`. As far as I'm aware, different methods
   see different parts of that, where typical dictionary operations
   ignore the presence of the newcomers, and no wonder since they are
   initialised after the `super.__init()__` has run and are not part of
   the original design.
 * due to the inheritance from the standard Python dict, many inherited
   methods - that are in and by themselves useful - have become
   untrustworthy. This extends beyond the above-mentioned "not seeing"
   part and includes these methods being seemingly hardcoded to dicts
   in some respect. For example, the inherited `.copy()` method returns
   a dictionary as output, which is very unhelpful. It therefore seems
   unsafe to trust that many other methods will behave more to our
   liking.

Next point of order after this commit should therefore be to restructure
the `Snapshot` class and make the dictionary just a part of it -
probably as its own attribute called `attributes` (or something less
common) that is a dict itself.

Resolves issue #20
Related to issue #19