Reworked snapshots in SiGMo - major changes!
The following changes were made over the course of the last months. This
and the previous/subsequent commits with "major changes!" in the title
are to serve as a new basis for the continued, more atomic development
of the code.

The `Snapshot` class was rewritten to be an independent class and not
be a subclass of the standard python dictionary anymore, eliminating
most of the questionable behaviour mentioned in issue #21 and indeed in
the commit 342437ef1e63b7e1bbcfdb2e9d7b99cd609a59f3 itself.

Beyond that, some methods are renamed to be more intuitive and
consistent, and some useful helper methods like `autoname_with_index`
are introduced.

All in all, this is a major commit that should have been split into
many, many atomic commits, but alas, it is how it is, and untangling the
major developments that occurred since the last commit is both
unfeasible and nonsensical, since major changes to the architecture of
the code were made.

Resolves issues #11, #21