Introduced init. conditions class - major changes!
The following additions were made in the course of the code's re-work
during the last months. This and the previous/subsequent commits with
"major changes!" in the title are to serve as a new basis for the
continued, more atomic development of the code.

Introduce a new way to build initial conditions (ICs) for SiGMo, using
the custom `SiGMo.IC` class. As opposed to the convoluted, very scripty
and grid-based IC set-up that was necessary before, this class allows
the user to create ICs for a number of n AstroObjects (of same type),
then certain add them to other m ICs for the same AstroObjects but
pertaining to other quantities, to create ICs that cover all n objects
with regard to their m properties.

Since the class defines sensible `__add__` and `__sub__` methods, the
adding and removing of ICs is intuitively possible. Schematics for
`__mult__` and `__truediv__` are also sketched out, but not supported
at the moment.

The method `__len__` returns the number of different parameter
configurations aka later-to-be AstroObjects, while `__getitem__`
enables unpacking.

The `allvalues` method returns all the values for one parameter, or if
several keys are provided, for several parameters.

The `create_AstroObject_Ensemble` method is a helper method to quickly
create 1 Environment, an arbitrary number of Halos and the same number
of Galaxies, with one Galaxy associated with each of the Halos.

All in all, this is a major commit that should have been split into
many, many atomic commits, but alas, it is how it is, and untangling the
major developments that occurred since the last commit is both
unfeasible and nonsensical, since major changes to the architecture of
the code were made.