Introduced integrator class(es) - major changes!
The following additions were made in the course of the code's re-work
during the last months. This and the previous/subsequent commits with
"major changes!" in the title are to serve as a new basis for the
continued, more atomic development of the code.

(1) We introduce an abstract base class called `Integrator` to serve as
    template for different types of integrators of our simulations.
    `Integrator` requires its subclasses to define their own `__init__`
    and `integrate` methods, and provides an (as of now unused) method
    to determine the length of the next time step called `determine_dt`.

(2) We also introduce two subclasses of `Integrator`:
    (a) the working `FTI` (Fixed Timestep Integrator) class, coming with
        its own set of methods;
    (b) the as of now empty `VTI` (Variable Timestep Integrator) class.

The integrators better capture the time integration and evolution
process that was before entirely written in the run scripts/main.py

All in all, this is a major commit that should have been split into
many, many atomic commits, but alas, it is how it is, and untangling the
major developments that occurred since the last commit is both
unfeasible and nonsensical, since major changes to the architecture of
the code were made.