Continuous change to the run script
A lot of the heavy lifting was taken out of this script and put into
for example the SiGMo_ICs.py and SiGMo_Integrator.py, in a more
structured and formalised way.

This script still carries most of the history of how to call the code
in many different, mostly outdated versions. It will need a major clean-
up at some point in the near future!