Implement the replacement in make_multi_snapshot
Implement the subsequent replacement of Halo and Galaxy entries in the
initially flat-ish (no AstroObjects) dict by flat-ish dicts of the
respective Halo or Galaxy objects. Depth depends on whether the self
object is an Environment (then depth 2: all Halos in the list, and all
Galaxies in each Halo's list), or a Halo (then depth 1: all Galaxies in
the Halo's list). For Galaxy objects, multi_snapshot is the same as
single_snapshot, since they are (as of now) lowest in hierarchy of
AstroObjects.