Include single/multi snapshot option in integrate
Explicitly included and then handed over, might think about using `args`
and `kwargs` in the future to streamline/slim down interfaces and reduce
the amount of changes necessary in other classes/routines due to changes
in one class employed. Here: changes in AstroObject necessitated changes
here, in Integrator, and the interface one was non-essential if all
arguments were just passed on and the necessary arguments popped from
the list/dict.