Add Jupyter Notebook for observ. vs sim.
This Notebook was developed off the Repo to just read in simulation
and (some summary) observation data, calculate SFR79, and plot/compare
different quantities over the runtime of the simulation as well as the
SFR79 from both observations and from our simulation(s).

Please bear in mind that Jupyter Notebooks are not great for git and
version control and the diffs might look terrible in GitHub, which is
the main reason I did not originally include the notebook (and some
others) in the repository.