Add observational SFR79 GMS notebook
This notebook (to the very largest part) predates the SiGMo code. It is
used to read in SDSS and (later-on) xCOLD GASS data, compute the obs.
SFR79 as well as other quantities, and plot/compare them in many
different ways.

This notebook also contains some useful definitions like Amelie's GMS
relation, Lilly's SFR79 fit function as well as the calibration
parameters for different redshifts, etc.

All of this is not strictly part of the code itself, but since I
routinely use SFR, M_star, SFR79 etc. from SDSS, and plan to use the
same and more from xCOLD GASS (also the gas masses) soon, it makes sense
to keep tabs on it.