Generate accretion scaled ICs acc. to GMS offset
Implemented initial conditions for our backwards simulation runs that
scale with the offset ΔSFR from the galaxy main sequence. It uses the
parameter `sMIR_scaling` and takes it to the power of the SFR offset
in dex:

`sMIR_scaling = (base factor)^ΔSFR`

This leads to no change in sMIR for a base factor of 1, and exponential
increase of accretion above/decrease below the main sequence. We use the
Leslie et al. (2020) GMS to keep this pice of IC setup code compatible
with starting redshifts other than z=0 (which is how we currently use
it).

Caveat: currently, the sMIR_scaling gets determined only at the initial
stage during IC generation; for a backward run, this corresponds to the
final environment/conditions of the galaxy, and might not be sensible
when trying to figure out where the galaxy came from, since this
effectively forces the final GMS position on the galaxy. We could
implement an (optional) update of `sMIR_scaling` during the run of the
simulation