De-couple snapshot numbering from timestep numbers
Previously, the number of the current integration time-step was
automatically used as the number of the snapshot file written to disk.
This caused problems when using any value for `wtd` (write to disk) than
1, so if any time steps were skipped in the outputs, since the read-in
assumes continuous numbering.

This problem has now been circumvented by independent, continuous
numbering of the outputs.