Deactivate afterwards calc of Lilly-like fractions
This is the only feature, which until now also went squarely unused,
that required the additional snapshots being generated at every step,
and them to be stored in the `previous` class attribute.

This has sped up the simulations by roughly 30-35 percent!

If the state of the object one timestep ago ever is needed again, these
lines can be un-commented