Add periodic fluctuation for sMIR_scaling
Added new function, analogous to the off-GMS scaling implemented a while
ago. From the docstring of the new function:

Specific function to compute a new value for the sMIR_scaling factor
which scales/regulates the Specific Mass Increase Rate with respect
to the (standard/commonly expected) value at a specific redshift.
periodic function, its period, and the physical period of the

This function varies sMIR_scaling periodically, based on an existing
fluctuation. The difference between the sMIR_scaling_basefactor and
the normal scaling (read: 1) is the amplitude of the resulting
fluctuation.

The resulting expression is
s = 1 + (b - 1) * f(c * t)  , with:
 * s : the `sMIR_scaling` factor that gets used in the computation
 of the halo MIR (the return value of this function), and
 * b : the `sMIR_scaling_basefactor` of the Halo, a value that is
   expected to be constant over the lifetime of the Halo instance
   and is only used for these kinds of calculations
 * f : the existing periodic function
 * c : the conversion factor from physical fluctuations to the
   period of the mathematical function
 * t : the time term, combining the current age of the halo and a
   fluctuation offset specific to that halo to avoid synchronous
   oscillations in ensembles of halos