Fix the update of age property
In all three subclasses of AstroObject, the `age` was not being updated
after the initial creation/setting to 0, neither by individually
incrementation nor by update from the parent object (e.g. for Halo and
Galaxy objects). This meant that methods like the one for periodic
fluctuations in accretion, that specifically rely on the age to
determine where in the current cycle they are, were not functioning
properly.

The issue has been fixed by incrementing the `AstroObject.age` at every
timestep in the respective `evolve` methods by the current `timestep`