Compare to MH2 relation as well as xCOLD GASS data
xCOLD GASS data in close enough proximity to our own simulated galaxies
is a bit sparse and apparently leads to sampling issues, with a jagged
effect on the Delta Mgas comparison. To assert that the fault doesn't
lie with the simulated galaxies, we now also include a way to compare
Mgas from sims with the MH2 relation from Saintong & Catinella 2022,
Equation 5:

log(Mmol/Mstar) = (0.75 ± 0.01) log(SSFR) + (6.24 ± 0.12) + log(1.36)

with the added last term/factor on the RHS for the total molecular mass
fraction instead of just the H2 fraction.

The flag used to turn this comarison on it `_useMgasRelation`.
`_bootstrapping` should be turned off, as is always uses the xCOLD GASS
values for comparison, no matter the other flags