Replace GMS from Saintonge+2016 by Saintonge+2022
Originally used a locally written version of the star-forming galaxy
main sequence from Saintonge+2016, which has the problem of a downturn
at high Mstar, which is assumed to be artificial. Saintonge+2022
provides an improved version of this relation, which has already been
implemented in `SiGMo_misc.py`. This is being used here, and should work
as a drop-in replacement. Results using or presenting the offset from
the GMS after this commit should differ somewhat from those calculated
before this change, albeit be more correct