Add comparison of sims with similar SDSS galaxies
Implement the SFR79 comparison that we had already done for xCOLD GASS
(there for both SFR79 and Mgas), but now with the much larger SDSS
sample. We can only do this with SFR79, as it does not depend on gas
detections, which are not available in SDSS - but in exchange, the
sample is a few orders of magnitude larger and should help us combat
statistical and samling scatter.

The Mgas comparison is made with the MH2 (and Mmol) relation from Eq. 5
of Saintonge & Catinella 2022