Include last intended time step in integrator
Until now, the primary time integration loop of `FTI.integrate()` did
only run from step t=1 to t=n_steps-1, due to the behaviour of range(),
or in this case tqdm.trange(), which mimics its behaviour.

This has been rectified, and as a result, results of simulation runs
might have very miniscule differences to before; standard time step size
at the moment is 1e-3 Gyr, with a runtime of this whole simulation of
approx. 10 Gyr